requirements.txt
Pillow==5.1.0
python-magic==0.4.15
